requirements.txt
flask[async]
discord.py
aiohttp
python-dotenv
gunicorn